dbx_app/requirements.txt
streamlit
databricks-sdk
openai
requests
pandas
